Select Box Drawing Properly, Interfaces
Put in an interface for RTS functions and drawing the selection box, moved the box to the HUD blueprint, instead of widget. Much easier.

As for selecting units with the box, I can probably do a function in the Interface, to help with that since I don't want the selected units, in the HUD