UI is now synced with which unit is selected
Units can add items to their inventory, and the display shows the inventory of whichever unit / squad leader is selected.

Although, when there is no unit selected, it keeps the inventory of the last selected unit displayed, which I'd like to fix. I've tried but couldn't figure it out exactly.