Players can now delete their save if they want to restart
I rescoped from my original plan of having a whole load and save manager menu, where you can select, delete, and load saves. As well as create new ones with custom names. That seems way out of scope and too complicated for the timeframe I have left. Couldn't figure out a simple way to go about it. So instead of committing to trying to do a sort of complicated system for me, I did this. It works